Add conversions between types
Add methods to Base.convert to convert between
Seis types with different parameters.

This is useful in case of setting fields (e.g., evt)
with objects with a different type parameter to
the field's type.

As per Base.convert, some or all of the converted
objects may alias the original.