Enable writing of CartTraces to SAC files
Previously, it was not posisble to write `CartTrace`s as SAC files
because it was assumed their events and stations contained
fields `lon`, `lat`, etc.  Relax this assumption so that `CartTrace`s
are written.

At present, the station and event coordinates are put in SAC
headers `USER0` to `USER5`, but this may change in future.