Finally fix plot annotations
Although we replaced the hacky call to `Main.Plots` for annotations
in df850f, there appears to be a bug in RecipesBase where
only the annotations in the final series in a recipe are plotted
(https://github.com/JuliaPlots/RecipesBase.jl/issues/81), so change
the order of plotting in the `plot` recipe so that both pick labels and
trace labels are properly shown.

Also take this opportunity to give a default x-axis label to trace
plots and tidy up the code to determine the limits of the whole
set of traces within the plotting window requested.