Improve comparison of angles for azimuths and inclinations
When comparing traces or stations (e.g., when using `is_north`,
`is_east`, and so on), a default angle tolerance is needed since
precision in these angles (which are often exactly 0° or 90°) may
be lost during calculations and subsequent writing.  Previously
we used `eps(T)`, but this is probably too small in practice, so
increase the tolerance to `√eps(T)`.

- To this end, add the function `_angle_tol` to return an appropriate
  tolerance for angle comparisons in °.
- Add an explanation of this to the `is_north`-type functions'
  docstrings.
- Make the `angle_difference` more tolerant of different numeric
  types by promoting carefully and avoiding `Float64` literals.