Change tolerance for angle comparisons
The previous default tolerance for angle comparisons in degrees
was `√eps(T)`, which for `Float64` is too small, leading to occasional
failures in testing of angles which should agree within error.
Increase the tolerance for `Float64`s.

Also add a multi-argument version of `Seis._angle_tol` so that
a suitable default tolerance can be picked when comparing
angles in different precisions.