Add are_orthogonal to test for orthogonality of traces (and stations)
`are_orthogonal` can take either two or three stations or traces
and determines if they are mutually orthogonal.  This works for traces
in any orientation and does not imply that they are horizontal or vertical
in particular.

The old internal function `Seis.traces_are_orthogonal` only tested
whether the component azimuths of two stations/traces were 90°
apart, which obviously does not work in general.  It will be removed
from a future version

Summary of changes:

- Add function `are_orthogonal`
- Add dependency on StaticArrays.
- Add helper functions:
  - `_direction_vector`
  - `_directions_are_orthogonal`
  - `_directions_are_parallel`
  - `_directions_are_antiparallel`
  - `_u_dot_v_and_theta`
- Add `TestHelpers` module to the `test` directory, containing
  helper functions for testing purposes.
- Add `src/compat.jl` to provide implementations of functionality
  not present in older supported Julia versions.