Read NPTS header if reading SAC file headers only
When reading SAC files with `read_sac(..; header_only=true)`,
fill in `.meta.SAC_npts` with NPTS as read from the file.  This
is because we usually calculate the number of points on the fly
in a `Trace` and so have no proper place to put it in a `Trace`.
However, when reading just the header and no trace data from a SAC
file, this information cannot be calculated and instead we need
to rely on the NPTS header value to know what this value should
be.

Alongside this change, also:
- Improve the `read_sac` docstring
- Fix a missing space in the error message when NPTS in a SAC
  file does not match the true trace length on disk
- Add a missing test for the correct behaviour of
  `Seis.SAC.read(file; check_npts=true)`