Refactor `cut[!]`
Replace the logic in `cut!` with helper function `_cut_time_indices`
which returns the cutting indices, new start time and whether or not the
new trace should be empty.

The idea behind this change is that people can implement their own
methods of `cut[!]` without having to reimplement the logic for finding
the current indices, but for now `_cut_time_indices` is not exported
and 'hidden' with an underscored name.

Divide up the `cut` tests into `@testsets`.