Implement trace merging with `merge[!]`
Allow sets of traces with the same channel code (or not if checking is
turned off) to be merged together.  This operation creates a new, single
trace containing the data of the previous separate traces.

The implementation is designed to be relatively efficient for multiple
traces (not just two) and only creates a single new array for the whole
new trace.  Benchmarking suggests performance is acceptable for `merge!`
while `merge` should be sped up as more copying is done than is needed.

Along the way:
- Add internal utility functions to enable snapping of samples to a
  common time base.
- Determine if traces are simply sequential with no gaps and thus allow
  for a fast path of just concatenating traces together.

Code is currently messy and only implemented for single-trace
`AbstractArray`s; any future trace array type (backed by matrix storage)
will require further work.