WIP: Add `plot_section[!]` to plot record sections with Makie
`plot_section` will plot a record section (like the existing
Plots-based `section`, but with the addition of logic to handle large
amounts of data so that data is redrawn at higher resolution when the
user zooms in to the data far enough (either interactively or with the
time limits set appropriately).  Otherwise, the default is to draw the
traces with min-max time binning (where each time point shows the
minimum and maximum value within that bin).

Add new helper functions (`_flatten_with_separator` and `_bin_min_max`)
to the `Plot` submodule to help implement the above.  They are added
here because they can be used for the Plots-based functionality in
future.

Some functions are not actually finished, so this commit should be
squashed into a future one.

Currently no trace filling is implemented.

No online documentation is yet written.