WIP: Stream filtering: add script with some example code
Begin the process of enabling stream filtering by adding a script
with example code.  By stream filtering, we mean being able to
filter multiple segments of sequential data with a stateful
filter, avoiding the need to read it all into memory.  This will
be useful to decimate large data volumes by applying a decimation
antialias filter as data is read and written.