Fix errors in overloading AbstractArray{<:Seis.AbstractData}
In the package, we add methods to `Base.getproperty` and
`Base.setproperty` for `Array`s of the types created here.  Up
until now, this has not caused any problems, even if it may be
type piracy.

However as of Julia v1.11, `Array` is now a Julia type and as such
`getproperty(::Array, ::Symbol)` is defined, which breaks the code
here.  Likewise, `setproperty!(::Array, ::Symbol)` is defined to
throw an error (since `Array`s should not have their fields
mutated).  See https://github.com/JuliaLang/julia/issues/56100
for details.

Update the existing `getproperty` and `setproperty!` methods and
add new ones for `Array` to overcome this problem.

Total test time does not appear to increase on Julia v1.10 with
this change.  (Julia v1.11 takes 2m45 for tests, v 1m45 for 1.10.)