Fix incorrect docstring for `differentiate`
Previously, the docstring incorrectly claimed that for `points=5`,
the number of points was reduced by 4 and the start time was
increased by twice the sampling interval.  In fact, the
implementation only removed two points and thus the start time
only increased by the sampling interval.  This is because for
`npoints=5` we use three-point differencing for the first and last
points of the new trace.  This is intentional, thus the
documentation was incorrect.