Add keyword argument constructors for Traces
Allow `Trace`s to be created with keyword arguments rather than
only positional arguments, and define defaults for the begin time,
sampling interval and trace content.

With `Trace`, `CartTrace` and `GeogTrace`, type parameters can be
specified as usual with the keyword argument constructors as well
as with the positional argument versions.