Use NanoDates for dates rather than DateTimes
Replace use of the Dates standard library type `DateTime` with
`NanoDate` of NanoDates.jl.  The means that dates are represented
internally with ns precision, and `NanoDate`s are returned when calling
e.g. `origin_time` or `dates`.

This also means that we no longer need to set `b` to anything other than
0 when reading miniSEED files (which are resolved usually to µs).