Use elevation not depth for Geographic and fix STEL IO for SAC files
Previously, the `Geographic <: Position` type had the field `dep` for
storing the vertical position of something in terms of its depth below
the reference level in km.  However, we dealt with this in an ad-hoc
way between `GeogEvent`s and `GeogStation`s, which both hold a
`Geographic` position as a field, because depth is more natural for an
event, but elevation is more natural for a station.  This meant that it
was necessary to convert between elevation and depth when constructing
stations.

Replace this ad-hoc approach by implementing automatic conversion
between elevation (in m) and depth (in km) via `getproperty` and
`setproperty!`, allowing access for both `.elev` and `.dep` fields for
`Geographic` objects.  `.elev` becomes the true field, which further
gives a sort of local right-handed coordinate system 'lon-lat-up'.

To make this more seamless, `propertynames` now lists `dep` for `Event`s
but `elev` for `Station`s, which simplifies the `show` methods.

Other changes along the way:
- Fix header name length calculation in `show`
- Improve docstrings for `Geographic` and `Cartesian`
- Test station elevation reading/writing for SAC files